Update Main.java
This is an update to Main.java. Threw out the setIdentifier method because currently, it didn't felt as if it fit within the code itself; went with a personality identifier, to which it would describe the person depending on if they're gloomy, irritable, timid or upbeat. Yes, it's based on Persona 5's demon negotiation types.